:art: Import cycles fixed! & update test with mocks used to circumvent import cycles
