+++ src/components/MapView/Legends/index.test.tsx
@@ -6,13 +6,6 @@ import Legends from '.';
 import { RootState, store } from '../../../context/store';
 
 jest.mock('./ColorIndicator', () => 'mock-ColorIndicator');
-jest.mock('../../../context/analysisResultStateSlice', () => ({
-  // if we try load the analysis redux slice we face an import cycle
-  /* eslint-disable no-unused-vars */
-  isAnalysisLayerActiveSelector: (_: RootState) => null,
-  analysisResultSelector: (_: RootState) => null,
-  /* eslint-enable no-unused-vars */
-}));
 
 test('renders as expected', () => {
   const { container } = render(
