+++ src/context/mapStateSlice/selectors.ts
@@ -5,6 +5,7 @@ import { Map as MapBoxMap } from 'mapbox-gl';
 import type { RootState } from '../store';
 import type { LayerDataTypes } from '../layers/layer-data';
 import type { MapState } from '.';
+import type { LayerKey } from '../../config/types';
 
 export const layersSelector = (state: RootState): MapState['layers'] =>
   state.mapState.layers;
@@ -12,8 +13,8 @@ export const dateRangeSelector = (state: RootState): MapState['dateRange'] =>
   state.mapState.dateRange;
 export const mapSelector = (state: RootState): MapBoxMap | undefined =>
   state.mapState.mapboxMap();
-// TODO: Improve the typing on this function with LayerKey from Analysis UI PR
-export const layerDataSelector = (id: string, date?: number) => (
+
+export const layerDataSelector = (id: LayerKey, date?: number) => (
   state: RootState,
 ): LayerDataTypes | undefined =>
   state.mapState.layersData.find(
